Add offset tracking to ReplicationClient
This commit adds an instance variable @offset to the ReplicationClient class in replication_client.rb. The offset is used to keep track of the current position in the
replication stream. When processing each command, the offset is updated based on the size of the encoded command and its arguments. This allows for accurate tracking of the
replication progress.